web-app/backend/requirements.txt
numpy==1.23.5
tensorflow==2.11.0
opencv-python==4.6.0.66

fastapi==0.85.0
uvicorn==0.19.0

python-dotenv==0.21.0
python-multipart==0.0.5